Fixes some compiler warnings
The warnings are related to:

* 64-to-32-bit conversion
* shadow variables

+++ cocos/2d/CCRenderTexture.cpp
@@ -198,8 +198,8 @@ bool RenderTexture::initWithWidthAndHeight(int w, int h, Texture2D::PixelFormat 
         glGetIntegerv(GL_FRAMEBUFFER_BINDING, &_oldFBO);
 
         // textures must be power of two squared
-        unsigned int powW = 0;
-        unsigned int powH = 0;
+        long powW = 0;
+        long powH = 0;
 
         if (Configuration::getInstance()->supportsNPOT())
         {
@@ -212,7 +212,7 @@ bool RenderTexture::initWithWidthAndHeight(int w, int h, Texture2D::PixelFormat 
             powH = ccNextPOT(h);
         }
 
-        int dataLen = (int)(powW * powH * 4);
+        long dataLen = (long)(powW * powH * 4);
         data = malloc(dataLen);
         CC_BREAK_IF(! data);
