+++ cocos/2d/ccCArray.cpp
@@ -28,10 +28,10 @@ THE SOFTWARE.
 
 NS_CC_BEGIN
 
-const int CC_INVALID_INDEX = -1;
+const long CC_INVALID_INDEX = -1;
 
 /** Allocates and initializes a new array with specified capacity */
-ccArray* ccArrayNew(int capacity)
+ccArray* ccArrayNew(long capacity)
 {
 	if (capacity == 0)
 		capacity = 7;
@@ -68,7 +68,7 @@ void ccArrayDoubleCapacity(ccArray *arr)
 	arr->arr = newArr;
 }
 
-void ccArrayEnsureExtraCapacity(ccArray *arr, int extra)
+void ccArrayEnsureExtraCapacity(ccArray *arr, long extra)
 {
 	while (arr->max < arr->num + extra)
     {
@@ -82,7 +82,7 @@ void ccArrayEnsureExtraCapacity(ccArray *arr, int extra)
 
 void ccArrayShrink(ccArray *arr)
 {
-    int newSize = 0;
+    long newSize = 0;
 	
 	//only resize when necessary
 	if (arr->max > arr->num && !(arr->num==0 && arr->max==1))
@@ -104,11 +104,11 @@ void ccArrayShrink(ccArray *arr)
 }
 
 /** Returns index of first occurrence of object, CC_INVALID_INDEX if object not found. */
-int ccArrayGetIndexOfObject(ccArray *arr, Object* object)
+long ccArrayGetIndexOfObject(ccArray *arr, Object* object)
 {
-    const int arrNum = arr->num;
+    const long arrNum = arr->num;
     Object** ptr = arr->arr;
-	for(int i = 0; i < arrNum; ++i, ++ptr)
+	for(long i = 0; i < arrNum; ++i, ++ptr)
     {
 		if( *ptr == object )
             return i;
@@ -143,7 +143,7 @@ void ccArrayAppendObjectWithResize(ccArray *arr, Object* object)
  enough capacity. */
 void ccArrayAppendArray(ccArray *arr, ccArray *plusArr)
 {
-	for(int i = 0; i < plusArr->num; i++)
+	for(long i = 0; i < plusArr->num; i++)
     {
 		ccArrayAppendObject(arr, plusArr->arr[i]);
     }
@@ -157,14 +157,14 @@ void ccArrayAppendArrayWithResize(ccArray *arr, ccArray *plusArr)
 }
 
 /** Inserts an object at index */
-void ccArrayInsertObjectAtIndex(ccArray *arr, Object* object, int index)
+void ccArrayInsertObjectAtIndex(ccArray *arr, Object* object, long index)
 {
 	CCASSERT(index<=arr->num, "Invalid index. Out of bounds");
 	CCASSERT(object != NULL, "Invalid parameter!");
 
 	ccArrayEnsureExtraCapacity(arr, 1);
 	
-	int remaining = arr->num - index;
+	long remaining = arr->num - index;
 	if( remaining > 0)
     {
 		memmove((void *)&arr->arr[index+1], (void *)&arr->arr[index], sizeof(Object*) * remaining );
@@ -176,7 +176,7 @@ void ccArrayInsertObjectAtIndex(ccArray *arr, Object* object, int index)
 }
 
 /** Swaps two objects */
-void ccArraySwapObjectsAtIndexes(ccArray *arr, int index1, int index2)
+void ccArraySwapObjectsAtIndexes(ccArray *arr, long index1, long index2)
 {
 	CCASSERT(index1>=0 && index1 < arr->num, "(1) Invalid index. Out of bounds");
 	CCASSERT(index2>=0 && index2 < arr->num, "(2) Invalid index. Out of bounds");
@@ -198,7 +198,7 @@ void ccArrayRemoveAllObjects(ccArray *arr)
 
 /** Removes object at specified index and pushes back all subsequent objects.
  Behavior undefined if index outside [0, num-1]. */
-void ccArrayRemoveObjectAtIndex(ccArray *arr, int index, bool bReleaseObj/* = true*/)
+void ccArrayRemoveObjectAtIndex(ccArray *arr, long index, bool bReleaseObj/* = true*/)
 {
     CCASSERT(arr && arr->num > 0 && index>=0 && index < arr->num, "Invalid index. Out of bounds");
     if (bReleaseObj)
@@ -208,7 +208,7 @@ void ccArrayRemoveObjectAtIndex(ccArray *arr, int index, bool bReleaseObj/* = tr
     
 	arr->num--;
 	
-	int remaining = arr->num - index;
+	long remaining = arr->num - index;
 	if(remaining>0)
     {
 		memmove((void *)&arr->arr[index], (void *)&arr->arr[index+1], remaining * sizeof(Object*));
@@ -218,16 +218,16 @@ void ccArrayRemoveObjectAtIndex(ccArray *arr, int index, bool bReleaseObj/* = tr
 /** Removes object at specified index and fills the gap with the last object,
  thereby avoiding the need to push back subsequent objects.
  Behavior undefined if index outside [0, num-1]. */
-void ccArrayFastRemoveObjectAtIndex(ccArray *arr, int index)
+void ccArrayFastRemoveObjectAtIndex(ccArray *arr, long index)
 {
 	CC_SAFE_RELEASE(arr->arr[index]);
-	int last = --arr->num;
+	long last = --arr->num;
 	arr->arr[index] = arr->arr[last];
 }
 
 void ccArrayFastRemoveObject(ccArray *arr, Object* object)
 {
-	int index = ccArrayGetIndexOfObject(arr, object);
+	long index = ccArrayGetIndexOfObject(arr, object);
 	if (index != CC_INVALID_INDEX)
     {
 		ccArrayFastRemoveObjectAtIndex(arr, index);
@@ -238,7 +238,7 @@ void ccArrayFastRemoveObject(ccArray *arr, Object* object)
  found the function has no effect. */
 void ccArrayRemoveObject(ccArray *arr, Object* object, bool bReleaseObj/* = true*/)
 {
-	int index = ccArrayGetIndexOfObject(arr, object);
+	long index = ccArrayGetIndexOfObject(arr, object);
 	if (index != CC_INVALID_INDEX)
     {
 		ccArrayRemoveObjectAtIndex(arr, index, bReleaseObj);
@@ -249,7 +249,7 @@ void ccArrayRemoveObject(ccArray *arr, Object* object, bool bReleaseObj/* = true
  first matching instance in arr will be removed. */
 void ccArrayRemoveArray(ccArray *arr, ccArray *minusArr)
 {
-	for(int i = 0; i < minusArr->num; i++)
+	for(long i = 0; i < minusArr->num; i++)
     {
 		ccArrayRemoveObject(arr, minusArr->arr[i]);
     }
@@ -259,8 +259,8 @@ void ccArrayRemoveArray(ccArray *arr, ccArray *minusArr)
  matching instances in arr will be removed. */
 void ccArrayFullRemoveArray(ccArray *arr, ccArray *minusArr)
 {
-	int back = 0;
-	int i = 0;
+	long back = 0;
+	long i = 0;
 	
 	for( i = 0; i < arr->num; i++) 
     {
@@ -282,7 +282,7 @@ void ccArrayFullRemoveArray(ccArray *arr, ccArray *minusArr)
 // #pragma mark ccCArray for Values (c structures)
 
 /** Allocates and initializes a new C array with specified capacity */
-ccCArray* ccCArrayNew(int capacity)
+ccCArray* ccCArrayNew(long capacity)
 {
 	if (capacity == 0)
     {
@@ -317,15 +317,15 @@ void ccCArrayDoubleCapacity(ccCArray *arr)
 }
 
 /** Increases array capacity such that max >= num + extra. */
-void ccCArrayEnsureExtraCapacity(ccCArray *arr, int extra)
+void ccCArrayEnsureExtraCapacity(ccCArray *arr, long extra)
 {
     ccArrayEnsureExtraCapacity((ccArray*)arr,extra);
 }
 
 /** Returns index of first occurrence of value, CC_INVALID_INDEX if value not found. */
-int ccCArrayGetIndexOfValue(ccCArray *arr, void* value)
+long ccCArrayGetIndexOfValue(ccCArray *arr, void* value)
 {
-	for( int i = 0; i < arr->num; i++)
+	for(long i = 0; i < arr->num; i++)
     {
 		if( arr->arr[i] == value )
             return i;
@@ -340,11 +340,11 @@ bool ccCArrayContainsValue(ccCArray *arr, void* value)
 }
 
 /** Inserts a value at a certain position. Behavior undefined if array doesn't have enough capacity */
-void ccCArrayInsertValueAtIndex( ccCArray *arr, void* value, int index)
+void ccCArrayInsertValueAtIndex( ccCArray *arr, void* value, long index)
 {
 	CCASSERT( index < arr->max, "ccCArrayInsertValueAtIndex: invalid index");
 	
-	int remaining = arr->num - index;
+	long remaining = arr->num - index;
     // make sure it has enough capacity
     if (arr->num + 1 == arr->max)
     {
@@ -385,7 +385,7 @@ void ccCArrayAppendValueWithResize(ccCArray *arr, void* value)
  enough capacity. */
 void ccCArrayAppendArray(ccCArray *arr, ccCArray *plusArr)
 {
-	for( int i = 0; i < plusArr->num; i++)
+	for( long i = 0; i < plusArr->num; i++)
     {
 		ccCArrayAppendValue(arr, plusArr->arr[i]);
     }
@@ -408,9 +408,9 @@ void ccCArrayRemoveAllValues(ccCArray *arr)
  Behavior undefined if index outside [0, num-1].
  @since v0.99.4
  */
-void ccCArrayRemoveValueAtIndex(ccCArray *arr, int index)
+void ccCArrayRemoveValueAtIndex(ccCArray *arr, long index)
 {
-	for( int last = --arr->num; index < last; index++)
+	for( long last = --arr->num; index < last; index++)
     {
 		arr->arr[index] = arr->arr[index + 1];
     }
@@ -421,9 +421,9 @@ void ccCArrayRemoveValueAtIndex(ccCArray *arr, int index)
  Behavior undefined if index outside [0, num-1].
  @since v0.99.4
  */
-void ccCArrayFastRemoveValueAtIndex(ccCArray *arr, int index)
+void ccCArrayFastRemoveValueAtIndex(ccCArray *arr, long index)
 {
-	int last = --arr->num;
+	long last = --arr->num;
 	arr->arr[index] = arr->arr[last];
 }
 
@@ -432,7 +432,7 @@ void ccCArrayFastRemoveValueAtIndex(ccCArray *arr, int index)
  */
 void ccCArrayRemoveValue(ccCArray *arr, void* value)
 {
-	int index = ccCArrayGetIndexOfValue(arr, value);
+	long index = ccCArrayGetIndexOfValue(arr, value);
 	if (index != CC_INVALID_INDEX)
     {
 		ccCArrayRemoveValueAtIndex(arr, index);
@@ -444,7 +444,7 @@ void ccCArrayRemoveValue(ccCArray *arr, void* value)
  */
 void ccCArrayRemoveArray(ccCArray *arr, ccCArray *minusArr)
 {
-	for(int i = 0; i < minusArr->num; i++)
+	for(long i = 0; i < minusArr->num; i++)
     {
 		ccCArrayRemoveValue(arr, minusArr->arr[i]);
     }
@@ -455,9 +455,9 @@ void ccCArrayRemoveArray(ccCArray *arr, ccCArray *minusArr)
  */
 void ccCArrayFullRemoveArray(ccCArray *arr, ccCArray *minusArr)
 {
-	int back = 0;
+	long back = 0;
 	
-	for(int i = 0; i < arr->num; i++)
+	for(long i = 0; i < arr->num; i++)
     {
 		if( ccCArrayContainsValue(minusArr, arr->arr[i]) ) 
         {
